feat: ゲームステージ、UI、タイトル画面の改善と拡張
詳細:

- **`StageGenerator`**
  - ステージ幅・高さの戻り値型を `int` から `double` に変更。
  - 新しいメソッド `getScreenObjects()` を追加、UIやアニメーションエレメントの管理を強化。

- **`RenderLayer`、`Team`**
  - 列挙型にコメントを追加し、コードの可読性と開発者理解を向上。

- **`Wall` クラス**
  - JavaDoc コメントをリファクタリングし、`@inheritDoc` を使用。コードの可読性を向上。

- **追加:** 新規クラス `TitleCharacter` を導入
  - タイトル画面のロゴアニメーションロジックを実装（テキストの回転、色変更、スケーリング）。

- **`GamePanel`、`GameEngine` の変更**
  - `GameEngine` に `ScreenObjects` を渡す機能を追加。
  - `GamePanel` でのパネル初期化ロジックを強化。

- **`GameWindow`、`StartPanel`**
  - `StartPanel` に「How to Play」ボタンを追加、ゲーム開始時の UX を向上。
  - `GameWindow` に `showHowToPlay` メソッドを追加（現在は TODO）。

+++ src/stage/GameStage.java
@@ -14,8 +14,8 @@ import static stage.Team.*;
 public class GameStage implements StageInfo {
 
 	// 描画範囲の情報
-	private final int stageWidth;
-	private final int stageHeight;
+	private final double stageWidth;
+	private final double stageHeight;
 
 	// オブジェクト管理
 	private int nextPrivateObjectID = 0;
@@ -47,17 +47,19 @@ public class GameStage implements StageInfo {
 
 	/**
 	 * ステージの幅を取得します。
+	 *
 	 * @return ステージの幅
 	 */
-	public int getStageWidth() {
+	public double getStageWidth() {
 		return stageWidth;
 	}
 
 	/**
 	 * ステージの高さを取得します。
+	 *
 	 * @return ステージの高さ
 	 */
-	public int getStageHeight() {
+	public double getStageHeight() {
 		return stageHeight;
 	}
